Add /healthcheck endpoint that doesn't collect PDU data
When the exporter is deployed in a system that healthchecks running
processes, such as Kubernetes, it's useful to have a healthcheck
endpoint that can return quickly which still catching errors where eg
the web server is hanging.

This basic /healthcheck endpoint returns a 200 and a short message,
avoiding the call to `collect` which would normally gather all of the
PDU data and can be slow to return. All other endpoints will still
return the PDU data in Prometheus format.

Closes
https://github.com/psyinfra/prometheus-raritan-pdu-exporter/issues/5.